WRR-12000: Enhanced ColorPicker code and made component public (#1759)
* added JS Docs

* removed @module and indented code

* fixed JS docs

* fixed `too many doclets` error

* fixed js docs warning

* fixed JS docs

* fixed js docs

* added JD docs for SpectrumIndicator.js

* refactored rgbStringToHex function and added `willReadFrequently: true` parameter in ColorPickerSpectrum for optimization

* made ColorPicker main component public and ColorPickerGrid component props private

* fixed qa sampler

* removed forgotten console.log

* removed transform as it conflicted with 5-way navigation

* review fixes

* modified example in JS docs

* modified JS docs for ColorPicker example

* JS docs fix

* fixed JS docs

* added default colors

* fixed typo

* fixed JS docs

* changed slider names and adjusted tabLayout width

* updated deprecated max-width value

* moved max-width properties to target tabs class

* corrected selector

* fixed typo

* added `type` and `disabled` props

* changed type prop to public

* added disabled style and behavior for each subcomponent of ColorPicker

* added missing dependency

* review fixes

* added missing dependency

* review fix

* WRR-3979: Updated storybook to 8.4 (#1739)

* updated storybook to 8.4

* updated to latest storybook

* updated storybook-utils

* review fix

* dummy commit to fix cla license

* revert dummy commit

* removed spotlightDisabled from Dropdown as it is not recognized as a custom prop

* undo spotlightDisabled removal - warning has different cause

---------
