+++ ColorPicker/ColorPicker.js
@@ -1,181 +1,244 @@
-/* eslint-disable react-hooks/rules-of-hooks */
 /**
- * Sandstone component to allow the user to choose a color.
+ * Sandstone component that allows the user to choose a color
+ * either from a grid, a spectrum, or RGB/HSL color sliders.
  *
  * @example
  * <ColorPicker
- *	 color={'#FF00FF'}
- *	 colorHandler={setColor}
- *	 presetColors={['#FF0000', '#00FF00', '#0000FF']}
- *	 text={'Color Picker'}
+ * 	 onChangeColor={() => console.log}
+ * 	 open
  * />
  *
  * @module sandstone/ColorPicker
  * @exports ColorPicker
  * @exports ColorPickerBase
- * @exports ColorPickerDecorator
- * @private
+ * @public
  */
-
-import kind from '@enact/core/kind';
 import Spottable from '@enact/spotlight/Spottable';
 import {Cell, Column, Row} from '@enact/ui/Layout';
-import Toggleable from '@enact/ui/Toggleable';
+import ri from '@enact/ui/resolution';
 import PropTypes from 'prop-types';
-import compose from 'ramda/src/compose';
-import {useCallback, useEffect, useState} from 'react';
+import {useCallback, useEffect, useRef, useState} from 'react';
 
-import BodyText from '../BodyText';
-import Button, {ButtonBase} from '../Button';
+import {ButtonBase} from '../Button';
 import Icon from '../Icon';
-import Item from '../Item';
 import Popup from '../Popup';
 import Skinnable from '../Skinnable';
-import Slider from '../Slider';
+import TabLayout, {Tab} from '../TabLayout';
 
-import {hexToHSL, HSLToHex} from './utils';
+import ColorPickerGrid from './ColorPickerGrid';
+import ColorPickerSlider from './ColorPickerSlider';
+import ColorPickerSpectrum from './ColorPickerSpectrum';
+import {generateOppositeColor} from './utils';
 
 import componentCss from './ColorPicker.module.less';
 
 const SpottableButton = Spottable(ButtonBase);
 
 /**
- * A component that contains the content for the {@link sandstone/ColorPicker|ColorPicker} popup.
+ * The favorite colors component.
  *
- * @class PopupContent
+ * This component is most often not used directly but may be composed within another component as it
+ * is within {@link sandstone/ColorPicker|ColorPicker}.
+ *
+ * @class FavoriteColors
  * @memberof sandstone/ColorPicker
  * @ui
  * @private
  */
-const PopupContent = ({color, colorHandler, css, presetColors}) => {
-	const [hue, setHue] = useState(0);
-	const [saturation, setSaturation] = useState(0);
-	const [lightness, setLightness] = useState(0);
+const FavoriteColors = ({disabled, favoriteColors = [], favoriteColorsHandler, selectedColor = '#3455eb', selectedColorHandler}) => {
+	const [clickEnabled, setClickEnabled] = useState(true);
+	const [editEnabled, setEditEnabled] = useState(false);
 
-	useEffect(() => {
-		let {h, s, l} = hexToHSL(color);
+	const shakeEffectRef = useRef(null);
+	const timerRef = useRef(null);
 
-		setHue(h);
-		setSaturation(s);
-		setLightness(l);
-	}, [color]);
+	const addNewFavoriteColor = useCallback(() => {
+		if (disabled) return;
+		if (favoriteColors.includes(selectedColor)) return;
+		favoriteColorsHandler(() => {
+			const colorsState = [...favoriteColors, selectedColor];
+			if (colorsState.length > 8) colorsState.shift();
 
-	const changeHue = useCallback((ev) => {
-		setHue(ev.value);
+			return colorsState;
+		});
+	}, [disabled, favoriteColors, favoriteColorsHandler, selectedColor]);
+
+	const onAddNewFavoriteColor = useCallback(() => {
+		if (disabled) return;
+		if (editEnabled) {
+			setEditEnabled(false);
+			return;
+		}
+		addNewFavoriteColor();
+	}, [addNewFavoriteColor, disabled, editEnabled]);
+
+	const onSelectFavoriteColor = useCallback((ev) => {
+		if (disabled) return;
+		if (!clickEnabled) return;
+		const targetId = ev.target.offsetParent.id || ev.target.id;
+		const [buttonColor, buttonIndex] = targetId.split('-');
+
+		if (editEnabled && clickEnabled) {
+			const filteredColors = favoriteColors.filter((color, index) => {
+				return 	!(color === buttonColor && index === Number(buttonIndex));
+			});
+
+			favoriteColorsHandler(filteredColors);
+			selectedColorHandler(selectedColor);
+			return;
+		}
+
+		favoriteColorsHandler(favoriteColors);
+		selectedColorHandler(buttonColor);
+	}, [clickEnabled, disabled, editEnabled, favoriteColors, favoriteColorsHandler, selectedColor, selectedColorHandler]);
+
+	const onPressHandler = useCallback((ev) => {
+		if (disabled) return;
+		if (editEnabled) return;
+		if (ev.type === 'pointerdown' || (ev.type === 'keydown' && ev.keyCode === 13)) {
+			const target = ev.target.id ? ev.target : ev.target.offsetParent;
+
+			shakeEffectRef.current = setTimeout(() => {
+				target.classList.add(componentCss.shakeFavoriteColor);
+			}, 300);
+
+			timerRef.current = setTimeout(() => {
+				setEditEnabled(true);
+				setClickEnabled(false);
+				target.classList.remove(componentCss.shakeFavoriteColor);
+			}, 1000);
+		}
+	}, [disabled, editEnabled]);
+
+	const onReleaseHandler = useCallback((ev) => {
+		const target = ev.target.id ? ev.target : ev.target.offsetParent;
+		target.classList.remove(componentCss.shakeFavoriteColor);
+
+		clearTimeout(shakeEffectRef.current);
+		clearTimeout(timerRef.current);
+		setTimeout(() => {
+			setClickEnabled(true);
+		}, 100);
 	}, []);
-
-	const changeLightness = useCallback((ev) => {
-		setLightness(ev.value);
-	}, []);
-
-	const changeSaturation = useCallback((ev) => {
-		setSaturation(ev.value);
-	}, []);
-
-	const handleClick = useCallback((ev) => {
-		colorHandler(ev.target.offsetParent.id);
-	}, [colorHandler]);
-
-	const onSliderValueChange = useCallback(() => {
-		colorHandler(HSLToHex(hue, saturation, lightness));
-	}, [colorHandler, hue, lightness, saturation]);
 
 	return (
-		<Cell className={css.colorPicker}>
-			<Row className={css.colorsRow} wrap>
-				{presetColors?.map((presetColor, presetColorIndex) => {
-
-					return (
-						<Cell key={presetColor + '-' + presetColorIndex} size="25%">
+		<div>
+			<Row className={componentCss.favoriteColorsRow}>
+				<Cell align="end">
+					{favoriteColors.slice(4, 8).map((color, index) => {
+						return (
 							<SpottableButton
-								className={css.coloredButton}
-								id={presetColor}
+								className={componentCss.favoriteColor}
+								id={`${color}-${index + 4}`}
+								key={`${color}_${index + 4}`}
 								minWidth={false}
-								onClick={handleClick}
-								style={{backgroundColor: presetColor}}
-								type="color"
-							/>
-						</Cell>
-					);
-				})}
+								onClick={onSelectFavoriteColor}
+								onKeyDown={onPressHandler}
+								onKeyUp={onReleaseHandler}
+								onPointerDown={onPressHandler}
+								onPointerUp={onReleaseHandler}
+								size="small"
+								spotlightDisabled={disabled}
+								style={{
+									backgroundColor: color,
+									borderColor: generateOppositeColor(color),
+									color: generateOppositeColor(color)
+								}}
+							>
+								{editEnabled && <Icon className={componentCss.deleteButton} size={'tiny'}>trash</Icon>}
+							</SpottableButton>
+						);
+					})}
+				</Cell>
+				<Cell align="end">
+					{favoriteColors.slice(0, 4).map((color, index) => {
+						return (
+							<SpottableButton
+								className={componentCss.favoriteColor}
+								id={`${color}-${index}`}
+								key={`${color}_${index}`}
+								minWidth={false}
+								onClick={onSelectFavoriteColor}
+								onKeyDown={onPressHandler}
+								onKeyUp={onReleaseHandler}
+								onPointerDown={onPressHandler}
+								onPointerUp={onReleaseHandler}
+								size="small"
+								spotlightDisabled={disabled}
+								style={{
+									backgroundColor: color,
+									borderColor: generateOppositeColor(color),
+									color: generateOppositeColor(color)
+								}}
+							>
+								{editEnabled && <Icon className={componentCss.deleteButton} size={'tiny'}>trash</Icon>}
+							</SpottableButton>
+						);
+					})}
+				</Cell>
 			</Row>
-			<div>
-				<Column className={css.colorPickerSliders}>
-					<BodyText className={css.colorSliderText} css={css}>Hue {hue}</BodyText>
-					<Slider
-						className={css.colorSlider}
-						max={356}
-						min={0}
-						onBlur={onSliderValueChange}
-						onClick={onSliderValueChange}
-						onChange={changeHue}
-						value={hue}
-					/>
-					<BodyText className={css.colorSliderText} css={css}>Saturation {saturation}%</BodyText>
-					<Slider
-						className={css.colorSlider}
-						max={100}
-						min={0}
-						onBlur={onSliderValueChange}
-						onClick={onSliderValueChange}
-						onChange={changeSaturation}
-						value={saturation}
-					/>
-					<BodyText className={css.colorSliderText} css={css}>Lightness {lightness}%</BodyText>
-					<Slider
-						className={css.colorSlider}
-						max={100}
-						min={0}
-						onBlur={onSliderValueChange}
-						onClick={onSliderValueChange}
-						onChange={changeLightness}
-						value={lightness}
-					/>
-				</Column>
-				<div className={css.coloredDiv} style={{backgroundColor: `hsl(${hue} ,${saturation}%, ${lightness}%)`}} />
-			</div>
-		</Cell>
+			<Column align="center" className={componentCss.selectedColorColumn}>
+				<SpottableButton
+					className={componentCss.selectedColor}
+					minWidth={false}
+					onClick={onAddNewFavoriteColor}
+					spotlightDisabled={disabled}
+					style={{
+						backgroundColor: selectedColor,
+						borderColor: generateOppositeColor(selectedColor),
+						color: generateOppositeColor(selectedColor)
+					}}
+				>
+					<Icon className={componentCss.selectedColorIcon} size="large">{editEnabled ? 'check' : 'plus'}</Icon>
+				</SpottableButton>
+			</Column>
+		</div>
 	);
 };
 
-PopupContent.propTypes = {
+FavoriteColors.displayName = 'FavoriteColors';
+
+FavoriteColors.propTypes = {
 	/**
-	 * Indicates the color.
+	 * Applies a disabled style and prevents interacting with the component.
 	 *
-	 * @type {String}
+	 * @type {Boolean}
+	 * @default false
 	 * @private
 	 */
-	color: PropTypes.string,
+	disabled: PropTypes.bool,
 
 	/**
-	 * Called when color is modified.
-	 *
-	 * @type {Function}
-	 * @private
-	 */
-	colorHandler: PropTypes.func,
-
-	/**
-	 * Customizes the component by mapping the supplied collection of CSS class names to the
-	 * corresponding internal elements and states of this component.
-	 *
-	 * The following classes are supported:
-	 *
-	 * `colorPicker` - The root class name
-	 * `coloredDiv`  - A class name used for a single div
-	 *
-	 * @type {Object}
-	 * @private
-	 */
-	css: PropTypes.object,
-
-	/**
-	 * Contains an array with a couple of possible preset colors.
+	 * Contains an array with the favorite colors.
 	 *
 	 * @type {Array}
 	 * @private
 	 */
-	presetColors: PropTypes.array
+	favoriteColors: PropTypes.array,
+
+	/**
+	 * Called when the favorite colors array is modified.
+	 *
+	 * @type {Function}
+	 * @private
+	 */
+	favoriteColorsHandler: PropTypes.func,
+
+	/**
+	 * Indicates the selected color.
+	 *
+	 * @type {String}
+	 * @private
+	 */
+	selectedColor: PropTypes.string,
+
+	/**
+	 * Called when the selected color is modified.
+	 *
+	 * @type {Function}
+	 * @private
+	 */
+	selectedColorHandler: PropTypes.func
 };
 
 /**
@@ -187,161 +250,157 @@ PopupContent.propTypes = {
  * @class ColorPickerBase
  * @memberof sandstone/ColorPicker
  * @ui
- * @private
+ * @public
  */
-const ColorPickerBase = kind({
-	name: 'ColorPicker',
+const ColorPickerBase = ({color = '#eb4034', colors = ['#eb4034', '#32a852', '#3455eb'], css, disabled, onChangeColor, open, type = 'grid', ...rest}) => {
+	const [favoriteColors, setFavoriteColors] = useState(colors);
+	const [selectedColor, setSelectedColor] = useState(color);
+	const [tabLayoutIndex, setTabLayoutIndex] = useState(0);
 
-	functional: true,
+	useEffect(() => {
+		setFavoriteColors(colors);
+		setSelectedColor(color);
 
-	propTypes: /** @lends sandstone/ColorPicker.ColorPickerBase.prototype */ {
-		/**
-		 * Indicates the color.
-		 *
-		 * @type {String}
-		 * @public
-		 */
-		color: PropTypes.string,
-
-		/**
-		 * Called when the color is modified.
-		 *
-		 * @type {Function}
-		 * @public
-		 */
-		colorHandler: PropTypes.func,
-
-		/**
-		 * Customizes the component by mapping the supplied collection of CSS class names to the
-		 * corresponding internal elements and states of this component.
-		 *
-		 * The following classes are supported:
-		 *
-		 * `colorPicker` - The root class name
-		 * `coloredDiv`  - A class name used for a single div
-		 *
-		 * @type {Object}
-		 * @public
-		 */
-		css: PropTypes.object,
-
-		/**
-		 * Applies the `disabled` class.
-		 *
-		 * When `true`, the color picker is shown as disabled.
-		 *
-		 * @type {Boolean}
-		 * @default false
-		 * @public
-		 */
-		disabled: PropTypes.bool,
-
-		/**
-		 * Called to open or close the color picker.
-		 *
-		 * @type {Function}
-		 * @public
-		 */
-		onTogglePopup: PropTypes.func,
-
-		/**
-		 * Indicates if the color picker is open.
-		 *
-		 * When `true`, contextual popup opens.
-		 *
-		 * @type {Boolean}
-		 * @default false
-		 * @private
-		 */
-		popupOpen: PropTypes.bool,
-
-		/**
-		 * Contains an array with a couple of possible preset colors.
-		 *
-		 * @type {Array}
-		 * @public
-		 */
-		presetColors: PropTypes.array,
-
-		/**
-		 * Contains the text that shows near color picker.
-		 *
-		 * @type {String}
-		 * @public
-		 */
-		text: PropTypes.string
-	},
-
-	handlers: {
-		handleClosePopup: (ev, {onTogglePopup}) => {
-			onTogglePopup();
-		},
-		handleOpenPopup: (ev, {disabled, onTogglePopup}) => {
-			if (!disabled) {
-				onTogglePopup();
-			}
+		switch (type) {
+			case 'grid':
+				setTabLayoutIndex(0);
+				return;
+			case 'spectrum':
+				setTabLayoutIndex(1);
+				return;
+			case 'sliders':
+				setTabLayoutIndex(2);
+				return;
+			default:
+				setTabLayoutIndex(0);
 		}
-	},
+	}, [color, colors, type]);
 
-	styles: {
-		css: componentCss,
-		publicClassNames: true
-	},
+	useEffect(() => {
+		if (disabled) return;
+		if (selectedColor || favoriteColors) {
+			onChangeColor({selectedColor, favoriteColors});
+		}
+	}, [disabled, favoriteColors, onChangeColor, selectedColor]);
 
-	render: ({color, colorHandler, css, disabled = false, handleClosePopup, handleOpenPopup, popupOpen = false, presetColors, text, ...rest}) => {
-		delete rest.onTogglePopup;
+	const handleGridClick = useCallback(() => {
+		if (disabled) return;
+		setTabLayoutIndex(0);
+	}, [disabled, setTabLayoutIndex]);
 
-		const CloseIcon = useCallback((props) => <Icon {...props} css={css} />, [css]);
-		const slotAfter = <SpottableButton
-			className={css.coloredButton}
-			disabled={disabled}
-			onClick={handleOpenPopup}
-			style={{backgroundColor: color}}
-			type="color"
-		/>;
+	const handleSpectrumClick = useCallback(() => {
+		if (disabled) return;
+		setTabLayoutIndex(1);
+	}, [disabled, setTabLayoutIndex]);
 
-		return (
-			<Cell shrink className={css.colorPicker}>
-				<Item disabled={disabled} onClick={handleOpenPopup} slotAfter={slotAfter} {...rest}>
-					{text}
-				</Item>
-				<Popup
-					className={css.colorPopup}
-					css={css}
-					noAnimation
-					onClose={handleClosePopup}
-					open={disabled ? false : popupOpen}
-					position="left"
-					scrimType="transparent"
-				>
-					<Row>
-						<Cell align="center">
-							<BodyText className={css.colorPopupHeader} css={css} noWrap>{text}</BodyText>
-						</Cell>
-						<Cell align="right" shrink>
-							<Button className={css.closeButton} css={css} iconComponent={CloseIcon} icon="closex" onClick={handleClosePopup} size="small" />
-						</Cell>
-					</Row>
-					<PopupContent color={color} colorHandler={colorHandler} css={css} presetColors={presetColors} />
-				</Popup>
-			</Cell>
-		);
-	}
-});
+	const handleSlidersClick = useCallback(() => {
+		if (disabled) return;
+		setTabLayoutIndex(2);
+	}, [disabled, setTabLayoutIndex]);
 
-/**
- * Applies Sandstone specific behaviors to {@link sandstone/ColorPicker.ColorPickerBase|ColorPicker} components.
- *
- * @hoc
- * @memberof sandstone/ColorPicker
- * @mixes sandstone/Skinnable.Skinnable
- * @mixes ui/Toggleable.Toggleable
- * @private
- */
-const ColorPickerDecorator = compose(
-	Skinnable,
-	Toggleable({prop: 'popupOpen', toggle: 'onTogglePopup'})
-);
+	return (
+		<Popup disabled={disabled} open={open} position="center" {...rest}>
+			<Row>
+				<Cell size="75%">
+					<TabLayout className={componentCss.pickerTabLayout} css={css} index={tabLayoutIndex} orientation="horizontal">
+						<Tab onTabClick={handleGridClick} spotlightDisabled={disabled} style={{width: ri.scaleToRem(400)}} title="Grid">
+							<div className={componentCss.colorPicker}>
+								<ColorPickerGrid disabled={disabled} selectedColorHandler={setSelectedColor} />
+							</div>
+						</Tab>
+						<Tab onTabClick={handleSpectrumClick} spotlightDisabled={disabled} style={{width: ri.scaleToRem(400)}} title="Spectrum">
+							<div className={componentCss.colorPicker}>
+								<ColorPickerSpectrum disabled={disabled} selectedColor={selectedColor} selectedColorHandler={setSelectedColor} />
+							</div>
+						</Tab>
+						<Tab onTabClick={handleSlidersClick} spotlightDisabled={disabled} style={{width: ri.scaleToRem(400)}} title="Sliders">
+							<div className={componentCss.colorPicker}>
+								<ColorPickerSlider disabled={disabled} selectedColor={selectedColor} selectedColorHandler={setSelectedColor} />
+							</div>
+						</Tab>
+					</TabLayout>
+				</Cell>
+				<Cell align="end" size="25%">
+					<Column>
+						<FavoriteColors
+							disabled={disabled}
+							favoriteColors={favoriteColors}
+							favoriteColorsHandler={setFavoriteColors}
+							selectedColor={selectedColor}
+							selectedColorHandler={setSelectedColor}
+						/>
+					</Column>
+				</Cell>
+			</Row>
+		</Popup>
+	);
+};
+
+ColorPickerBase.displayName = 'ColorPicker';
+
+ColorPickerBase.propTypes = {/** @lends sandstone/ColorPicker.ColorPickerBase.prototype */
+	/**
+	 * Indicates the selected color.
+	 *
+	 * @type {String}
+	 * @public
+	 */
+	color: PropTypes.string,
+
+	/**
+	 * Contains an array with colors.
+	 *
+	 * @type {Array}
+	 * @public
+	 */
+	colors: PropTypes.array,
+
+	/**
+	 * Customizes the component by mapping the supplied collection of CSS class names to the
+	 * corresponding internal elements and states of this component.
+	 *
+	 * @type {Object}
+	 * @public
+	 */
+	css: PropTypes.object,
+
+	/**
+	 * Applies a disabled style and prevents interacting with the component.
+	 *
+	 * @type {Boolean}
+	 * @default false
+	 * @public
+	 */
+	disabled: PropTypes.bool,
+
+	/**
+	 * Called when the selected color is modified.
+	 *
+	 * @type {Function}
+	 * @public
+	 */
+	onChangeColor: PropTypes.func,
+
+	/**
+	 * Controls the visibility of the ColorPicker.
+	 *
+	 * By default, the ColorPicker and its contents are not rendered until `open`.
+	 *
+	 * @type {Boolean}
+	 * @default false
+	 * @public
+	 */
+	open: PropTypes.bool,
+
+	/**
+	 * Set the type of color picker to use.
+	 *
+	 * @type {('grid'|'spectrum'|'sliders')}
+	 * @default 'grid'
+	 * @public
+	 */
+	type: PropTypes.oneOf(['grid', 'spectrum', 'sliders'])
+};
 
 /**
  * A color picker component, ready to use in Sandstone applications.
@@ -349,15 +408,13 @@ const ColorPickerDecorator = compose(
  * @class ColorPicker
  * @memberof sandstone/ColorPicker
  * @extends sandstone/ColorPicker.ColorPickerBase
- * @mixes sandstone/ColorPicker.ColorPickerDecorator
  * @ui
- * @private
+ * @public
  */
-const ColorPicker = ColorPickerDecorator(ColorPickerBase);
+const ColorPicker = Skinnable(ColorPickerBase);
 
 export default ColorPicker;
 export {
 	ColorPicker,
-	ColorPickerBase,
-	ColorPickerDecorator
+	ColorPickerBase
 };
