+++ ColorPicker/ColorPickerSpectrum.module.less
