+++ ColorPicker/package.json
@@ -1,3 +1,3 @@
 {
-	"main": "ColorPicker.js"
+  "main": "ColorPicker.js"
 }
