--- ColorPickerPOC/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "ColorPickerPOC.js"
-}
