+++ ColorPickerSettingsApp/ColorPickerSettingsApp.module.less
@@ -1,0 +1,93 @@
+// ColorPickerSettingsApp.module.less
+//
+@import '../styles/colors.less';
+@import '../styles/mixins.less';
+@import '../styles/variables.less';
+
+.colorPickerSettingsApp {
+	align-items: center;
+	justify-content: space-between;
+
+	.colorsRow {
+		width: 522px;
+		padding-top: 9px;
+	}
+
+	.coloredButton {
+		border-radius: 72px;
+		border: 3px solid #ffffff;
+		height: 120px;
+		margin-left: 0;
+		min-width: 120px;
+		padding: 0;
+		transform: none;
+
+		.focus({
+			transform: scale(1.1);
+		});
+
+		.disabled({
+			opacity: @sand-disabled-opacity;
+		})
+	}
+
+	.coloredDiv,
+	.coloredInput {
+		border-radius: 24px;
+	}
+
+	.coloredDiv {
+		display: block;
+		height: 81px;
+		border: 3px solid #ffffff;
+
+		&.button, &.focusExpand {
+			margin: 24px;
+		}
+
+		.focus({
+			transform: scale(1.1);
+		});
+	}
+
+	.coloredInput {
+		opacity: 0;
+
+		&:hover {
+			cursor: pointer;
+		}
+	}
+}
+
+.colorPopup {
+	height: max-content;
+
+	.colorPopupHeader {
+		font-size: 66px;
+		text-transform: capitalize;
+		margin: 0 0 0 15px;
+	}
+
+	.closeButton {
+		&.button, &.focusExpand {
+			margin: 0;
+		}
+
+		.icon {
+			font-size: 48px;
+		}
+	}
+
+	.colorPickerSliders {
+		padding: 45px 0;
+
+		.colorSlider {
+			margin: 0 27px;
+		}
+
+		.colorSliderText {
+			font-size: 54px;
+			margin: 0 24px;
+		}
+	}
+}
